robot/requirements.txt
absl-py==2.1.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.3.0
blinker==1.9.0
click==8.1.8
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
Flask==3.0.3
fonttools==4.55.3
freetype-py==2.5.1
frozenlist==1.5.0
fsspec==2024.12.0
grpcio==1.68.1
hydra-core==1.3.2
imageio==2.36.1
importlib_metadata==8.5.0
ipython==8.31.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lightning==2.5.0.post0
lightning-utilities==0.11.9
lxml==5.3.0
Markdown==3.7
matplotlib==3.10.0
matplotlib-inline==0.1.7
mkl-service==2.4.0
multidict==6.1.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8
numpy==1.26.4
omegaconf==2.3.0
open3d==0.18.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
propcache==0.2.1
protobuf==5.29.2
ptyprocess==0.7.0
pure_eval==0.2.3
pycollada==0.6
pyglet==2.0.20
Pygments==2.18.0
PyOpenGL==3.1.0
pyparsing==3.2.0
pyquaternion==0.9.9
pyrender==0.1.45
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2024.2
referencing==0.35.1
retrying==1.3.4
rpds-py==0.22.3
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
stack-data==0.6.3
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
torchvision==0.20.1
tqdm==4.67.1
traitlets==5.14.3
trimesh==4.5.3
triton==3.1.0
tzdata==2024.2
urdfpy==0.0.22
wcwidth==0.2.13
Werkzeug==3.0.6
widgetsnbextension==4.0.13
yarl==1.18.3
zipp==3.21.0
